--- a/.betterer.log
+++ b/.betterer.log
@@ -1,0 +1,393 @@
+
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+313ms | 313ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+1ms | 1ms | Starting betterer
+37ms | 37ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+35ms | 35ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+34ms | 34ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+36ms | 36ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+34ms | 34ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+271ms | 271ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+263ms | 263ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+278ms | 278ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+270ms | 270ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+36ms | 36ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+37ms | 37ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+38ms | 38ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+1ms | 1ms | Starting eslint
+0ms | 0ms | Starting betterer
+37ms | 36ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+50ms | 50ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+35ms | 35ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+1ms | 1ms | Starting eslint
+35ms | 34ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+1ms | 1ms | Starting eslint
+0ms | 0ms | Starting betterer
+35ms | 34ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+38ms | 38ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+35ms | 35ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+37ms | 37ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+1ms | 1ms | Starting eslint
+38ms | 37ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+40ms | 40ms | Inside eslint's test function
+1ms | 1ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+36ms | 36ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+1ms | 1ms | Starting eslint
+35ms | 34ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+39ms | 39ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+35ms | 35ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+36ms | 36ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+36ms | 36ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+36ms | 36ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+1ms | 1ms | Starting eslint
+36ms | 35ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+37ms | 37ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+38ms | 38ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+1ms | 1ms | Starting eslint
+37ms | 36ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+38ms | 38ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+38ms | 38ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+34ms | 34ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+1ms | 1ms | Starting eslint
+0ms | 0ms | Starting betterer
+37ms | 36ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+36ms | 36ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+1ms | 1ms | Starting eslint
+0ms | 0ms | Starting betterer
+36ms | 35ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+1ms | 1ms | Starting eslint
+0ms | 0ms | Starting betterer
+35ms | 34ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+1ms | 1ms | Starting eslint
+1ms | 1ms | Starting betterer
+41ms | 40ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+35ms | 35ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+35ms | 35ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+34ms | 34ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+35ms | 35ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+1ms | 1ms | Starting eslint
+0ms | 0ms | Starting betterer
+37ms | 36ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+38ms | 38ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+44ms | 44ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+38ms | 38ms | Inside eslint's test function
+1ms | 1ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+37ms | 37ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+35ms | 35ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+35ms | 35ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+1ms | 1ms | Starting eslint
+0ms | 0ms | Starting betterer
+37ms | 36ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+39ms | 39ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+1ms | 1ms | Starting eslint
+0ms | 0ms | Starting betterer
+38ms | 37ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+35ms | 35ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+35ms | 35ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+40ms | 40ms | Inside eslint's test function
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+35ms | 35ms | Inside eslint's test function
+167ms | 132ms | How many groups? 1
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+37ms | 37ms | Inside eslint's test function
+169ms | 132ms | How many groups? 1
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+37ms | 37ms | Inside eslint's test function
+173ms | 136ms | How many groups? 1
+1450ms | 1277ms | Done with eslint
+1ms | 1ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+37ms | 37ms | Inside eslint's test function
+167ms | 130ms | How many groups? 1
+1444ms | 1277ms | Done with eslint
+0ms | 0ms | Starting betterer
+1ms | 1ms | Starting betterer
+1ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+38ms | 37ms | Inside eslint's test function
+172ms | 134ms | How many groups? 1
+1372ms | 1200ms | Finished eslint
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+36ms | 36ms | Inside eslint's test function
+167ms | 131ms | How many groups? 1
+1377ms | 1210ms | Finished eslint
+1ms | 1ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+38ms | 38ms | Inside eslint's test function
+159ms | 121ms | How many groups? 1
+1374ms | 1215ms | Finished eslint
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+35ms | 35ms | Inside eslint's test function
+156ms | 121ms | How many groups? 1
+1372ms | 1216ms | Finished eslint
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+34ms | 34ms | Inside eslint's test function
+154ms | 120ms | How many groups? 1
+1369ms | 1215ms | Finished eslint
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+33ms | 33ms | Inside eslint's test function
+153ms | 120ms | How many groups? 1
+1355ms | 1202ms | Finished eslint
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+34ms | 34ms | Inside eslint's test function
+154ms | 120ms | How many groups? 1
+1358ms | 1204ms | Finished eslint
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+36ms | 36ms | Inside eslint's test function
+1405ms | 1248ms | Finished eslint
+157ms | 121ms | How many groups? 1
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+34ms | 34ms | Inside eslint's test function
+167ms | 133ms | How many groups? 1
+1382ms | 1215ms | Finished eslint
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+34ms | 34ms | Inside eslint's test function
+169ms | 135ms | How many groups? 1
+1409ms | 1240ms | Finished eslint
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting betterer
+0ms | 0ms | Starting eslint
+33ms | 33ms | Inside eslint's test function
+164ms | 131ms | How many groups? 1
+1396ms | 1232ms | Finished eslint