**What is this feature?**

Modifies the betterer configuration file to try to speed up pre-commit hooks

#### How faster is this?

Around 300ms faster. This is the difference between the `find` command (in unix systems) and the `glob` that comes with betterer.

Without this PR:

```❯ hyperfine 'yarn betterer precommit packages/grafana-ui/src/components/Tabs/Counter.tsx packages/grafana-ui/src/options/builder/text.tsx' --runs 10
Benchmark 1: yarn betterer precommit packages/grafana-ui/src/components/Tabs/Counter.tsx packages/grafana-ui/src/options/builder/text.tsx
  Time (mean ± σ):      4.813 s ±  0.064 s    [User: 13.855 s, System: 2.082 s]
  Range (min … max):    4.774 s …  4.990 s    10 runs

```

With this PR:

```❯ hyperfine 'yarn betterer precommit packages/grafana-ui/src/components/Tabs/Counter.tsx packages/grafana-ui/src/options/builder/text.tsx' --runs 10
Benchmark 1: yarn betterer precommit packages/grafana-ui/src/components/Tabs/Counter.tsx packages/grafana-ui/src/options/builder/text.tsx
  Time (mean ± σ):      4.538 s ±  0.105 s    [User: 13.581 s, System: 2.287 s]
  Range (min … max):    4.429 s …  4.793 s    10 runs
```

**Why do we need this feature?**

Currently betterer takes a long chunk of time in pre-commit hooks
**Who is this feature for?**

[Add information on what kind of user the feature is for.]

**Which issue(s) does this PR fix?**:

<!--

- Automatically closes linked issue when the Pull Request is merged.

Usage: "Fixes #<issue number>", or "Fixes (paste link of issue)"

-->

Fixes #

**Special notes for your reviewer:**

Please check that:

- [ ] It works as expected from a user's perspective.
- [ ] If this is a pre-GA feature, it is behind a feature toggle.
- [ ] The docs are updated, and if this is a [notable improvement](https://grafana.com/docs/writers-toolkit/writing-guide/contribute-release-notes/#how-to-determine-if-content-belongs-in-a-whats-new-document), it's added to our [What's New](https://grafana.com/docs/writers-toolkit/writing-guide/contribute-release-notes/) doc.